requirements.txt
streamlit==1.31.0
streamlit-webrtc==0.47.1  # Check for the latest compatible version
torch==2.2.2
torchaudio==2.2.2
torchvision==0.21.0
transformers==4.34.0
diffusers==0.21.4
accelerate==0.24.1
librosa==0.10.1
soundfile==0.12.1
matplotlib==3.7.2
numpy==1.24.3
Pillow==10.0.0
scipy==1.11.2
huggingface_hub==0.18.0
python-dateutil==2.8.2
av==10.0.0
aiortc==1.5.0
pydub==0.25.1
